Remove `iati create-database`; Add confirmation prompt to `iati drop-database`

+++ iati_datastore/iatilib/console.py
@@ -6,6 +6,7 @@ import subprocess
 
 import click
 from flask.cli import FlaskGroup, with_appcontext
+import flask_migrate
 import requests
 from sqlalchemy import not_
 
@@ -82,11 +83,7 @@ def parse_file(filenames, verbose=False, fail_xml=False, fail_spec=False):
 
 @cli.command()
 @with_appcontext
-def create_database():
-    db.create_all()
-
-
-@cli.command()
-@with_appcontext
 def drop_database():
+    """Drop all database tables."""
+    click.confirm('Are you sure?', abort=True)
     db.drop_all()
